Implemented review changes - style & Target options -> std::vector
