+++ include/tvm/build_module.h
@@ -7,7 +7,6 @@
 #define TVM_BUILD_MODULE_H_
 
 #include <string>
-#include "./tvm/c_dsl_api.h"
 #include "./tvm/runtime/packed_func.h"
 #include "./tvm/schedule_pass.h"
 #include "./tvm/lowered_func.h"
@@ -16,44 +15,44 @@ namespace tvm {
 
 /*!
 * \brief Container for target device information.
-* Use target_llvm, target_cuda etc functions instead of constructing directly.
+* Use target::llvm, target::cuda etc functions instead of constructing directly.
 */
 struct Target {
-    /*! \brief The name of the target device */
-    std::string target_name;
-    /*! \brief The type of the target device */
-    DLDeviceType device_type;
-    /*! \brief The maximum threads that a schedule should use for this device */
-    int max_num_threads = 1;
-    /*! \brief The warp size that should be used by the LowerThreadAllreduce pass */
-    int thread_warp_size = 1;
-    /*! \brief Keys for this target */
-    std::unordered_set<std::string> keys;
-    /*! \brief Options for this target */
-    std::unordered_set<std::string> options;
+  /*! \brief The name of the target device */
+  std::string target_name;
+  /*! \brief The type of the target device */
+  DLDeviceType device_type;
+  /*! \brief The maximum threads that a schedule should use for this device */
+  int max_num_threads = 1;
+  /*! \brief The warp size that should be used by the LowerThreadAllreduce pass */
+  int thread_warp_size = 1;
+  /*! \brief Keys for this target */
+  std::unordered_set<std::string> keys;
+  /*! \brief Options for this target */
+  std::vector<std::string> options;
 
+  Target(const std::string& target_name, DLDeviceType device_type, int max_num_threads,
+    int thread_warp_size, const std::unordered_set<std::string>& keys,
+    const std::vector<std::string>& options) :
+    target_name(target_name),
+    device_type(device_type),
+    max_num_threads(max_num_threads),
+    thread_warp_size(thread_warp_size),
+    keys(keys),
+    options(options) {
+  }
 
-    Target(const std::string& target_name, DLDeviceType device_type, int max_num_threads,
-        int thread_warp_size, const std::unordered_set<std::string>& keys,
-        const std::unordered_set<std::string>& options) {
-        this->target_name = target_name;
-        this->device_type = device_type;
-        this->max_num_threads = max_num_threads;
-        this->thread_warp_size = thread_warp_size;
-        this->keys = keys;
-        this->options = options;
-    }
+  /*! \return the full device string to pass to codegen::Build */
+  EXPORT std::string str() const;
 
-    /*! \return the full device string to pass to codegen::Build */
-    EXPORT std::string str() const;
-
-    /*!
-     * \brief Create a Target given a string
-     * \param target_str the string to parse
-     */
-    EXPORT static Target create(const std::string& target_str);
+  /*!
+   * \brief Create a Target given a string
+   * \param target_str the string to parse
+   */
+  EXPORT static Target create(const std::string& target_str);
 };
 
+/*! \brief This namespace provides functions to construct Target instances */
 namespace target {
 /*! \return A target for LLVM */
 EXPORT Target llvm();
@@ -75,50 +74,47 @@ EXPORT Target stackvm();
 
 }  // namespace target
 
-
 /*!
 * \brief Container for build configuration options
 */
 struct BuildConfig {
-    /*!
-     * \brief The data alignment to use when constructing buffers. If this is set to
-     * -1, then TVM's internal default will be used
-     */
-    int data_alignment = -1;
-    /*!
-     * \brief The offset factor to use when constructing buffers. If this is set to
-     * 0, then the offset field is not used.
-     */
-    int offset_factor = 0;
+  /*!
+   * \brief The data alignment to use when constructing buffers. If this is set to
+   * -1, then TVM's internal default will be used
+   */
+  int data_alignment = -1;
+  /*!
+   * \brief The offset factor to use when constructing buffers. If this is set to
+   * 0, then the offset field is not used.
+   */
+  int offset_factor = 0;
 
-    /*!
-     * \brief Splitting factor for loop splitting. If this is set to zero, no splitting will be
-     * done. Otherwise, a split will be done with this factor and the inner loop will be unrolled.
-     */
-    int double_buffer_split_loop = 1;
-    /*! \brief Threshold of number of steps in the loop to be automatically unrolled */
-    int auto_unroll_max_step = 0;
-    /*! \brief The maximum nested level of loops that can be automatically unrolled */
-    int auto_unroll_max_depth = 8;
-    /*! \brief The maximum extent of loop that will be unrolled */
-    int auto_unroll_max_extent = 0;
-    /*!
-     * \brief Whether to explicitly unroll the loop. If set to false, the unroll hint will
-     * be passed to the CodeGen phase. Set to true if CodeGen supports unroll pragma.
-     */
-    bool unroll_explicit = true;
+  /*!
+   * \brief Splitting factor for loop splitting. If this is set to zero, no splitting will be
+   * done. Otherwise, a split will be done with this factor and the inner loop will be unrolled.
+   */
+  int double_buffer_split_loop = 1;
+  /*! \brief Threshold of number of steps in the loop to be automatically unrolled */
+  int auto_unroll_max_step = 0;
+  /*! \brief The maximum nested level of loops that can be automatically unrolled */
+  int auto_unroll_max_depth = 8;
+  /*! \brief The maximum extent of loop that will be unrolled */
+  int auto_unroll_max_extent = 0;
+  /*!
+   * \brief Whether to explicitly unroll the loop. If set to false, the unroll hint will
+   * be passed to the CodeGen phase. Set to true if CodeGen supports unroll pragma.
+   */
+  bool unroll_explicit = true;
 
-    /*! \brief Set to true if buffer arguments do not overlap. This enables more optimization. */
-    bool restricted_func = true;
+  /*! \brief Set to true if buffer arguments do not overlap. This enables more optimization. */
+  bool restricted_func = true;
 
-    /*! \brief Whether to detect global barrier */
-    bool detect_global_barrier = false;
+  /*! \brief Whether to detect global barrier */
+  bool detect_global_barrier = false;
 
-    BuildConfig() {
-    }
+  BuildConfig() {
+  }
 };
-
-
 
 /*!
 * \brief Build a LoweredFunc given a schedule, args and binds
@@ -130,7 +126,7 @@ struct BuildConfig {
 * \return The lowered function.
 */
 EXPORT Array<LoweredFunc> lower(Schedule sch, const Array<Tensor>& args, const std::string& name,
-    const std::unordered_map<Tensor, Buffer>& binds, const BuildConfig& config);
+  const std::unordered_map<Tensor, Buffer>& binds, const BuildConfig& config);
 
 /*!
 * \brief Build a device and host module for a specific target from an array of lowered functions.
@@ -141,7 +137,7 @@ EXPORT Array<LoweredFunc> lower(Schedule sch, const Array<Tensor>& args, const s
 * \return The built module.
 */
 EXPORT runtime::Module build(const Array<LoweredFunc>& funcs, const Target& target,
-    Target* target_host, const BuildConfig& config);
+  Target* target_host, const BuildConfig& config);
 
 }  // namespace tvm
